third_party/WebKit/LayoutTests/http/tests/serviceworker/ready-controlled-document.html

<!DOCTYPE html>
<title>Service Worker: ready in a controlled document</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<body>
<script>
var t = async_test('ready in a controlled document');
t.step(function() {
    var url = 'resources/worker-no-op.js';
    var scope = 'resources/blank.html';
    service_worker_unregister_and_register(t, url, scope, onRegister);

    function onRegister(worker) {
        worker.addEventListener('statechange', t.step_func(function(event) {
            if (event.target.state == 'active')
                onActive();
        }));
    }

    var controller;

    function onActive() {
        with_iframe(scope, t.step_func(function(frame) {
            var w = frame.contentWindow;
            controller = w.navigator.serviceWorker.controller;
            w.navigator.serviceWorker.ready.then(t.step_func(onReady));
        }));
    }

    function onReady(serviceWorker) {
        assert_equals(serviceWorker, controller,
                      'ready should resolve to the controller of a ' +
                      'controlled document');
        service_worker_unregister_and_done(t, scope);
    }
});
</script>
